convert to multi command project
